make solutions return long.

+++ src/Day.kt
@@ -1,11 +1,11 @@
 import java.io.File
 
-abstract class Day(val test: Boolean = false, val expected1: Int = 0, val expected2: Int = 0) {
+abstract class Day(val test: Boolean = false, val expected1: Long = 0, val expected2: Long = 0) {
     val input: List<String> = File("src", "${this::class.simpleName}${if (test) "_test" else ""}.txt")
         .readLines()
 
-    abstract fun part1(): Int
-    abstract fun part2(): Int
+    abstract fun part1(): Long
+    abstract fun part2(): Long
     fun run() {
         println(part1())
         println(part2())
