+++ src/Day02.kt
@@ -1,10 +1,10 @@
 class Day02: Day() {
     val preparedInput = input.map {
         val split = it.split(" ")
-        Pair(split[0], split[1].toInt())
+        Pair(split[0], split[1].toLong())
     }
 
-    override fun part1(): Int {
+    override fun part1(): Long {
         val hor = preparedInput.filter { "forward".equals(it.first) }
             .map {it.second }.sum()
         val depth = preparedInput.filter { "down".equals(it.first) }
@@ -13,15 +13,15 @@ class Day02: Day() {
         return hor * depth
     }
 
-    override fun part2(): Int {
-        val aims = preparedInput.runningFold(0) {
+    override fun part2(): Long {
+        val aims = preparedInput.runningFold(0L) {
                 aim,
                 pair ->
             if ("down".equals(pair.first)) aim + pair.second
             else if ("up".equals(pair.first)) aim - pair.second
             else aim
         }
-        val depthHor = preparedInput.foldIndexed(Pair(0,0)) {
+        val depthHor = preparedInput.foldIndexed(Pair(0L,0L)) {
                 index, depthHor, currentInput ->
             if("forward".equals(currentInput.first))
                 Pair(depthHor.first + aims[index]*currentInput.second,
